Rename service files to appropriate names
